+++ docs/conf.py
@@ -46,10 +46,10 @@ html_theme_options = {
     "show_nav_level": 2,
     "navigation_depth": 2,
     "show_toc_level": 2,
-    # "external_links": [
-    #     {"name": "GitHub", "url": "https://github.com/corbel-spatial/ouroboros"},
-    #     {"name": "PyPI", "url": "https://pypi.org/project/ouroboros-gis/"},
-    # ],
+    "external_links": [
+        {"name": "GitHub", "url": "https://github.com/corbel-spatial/ouroboros"},
+        # {"name": "PyPI", "url": "https://pypi.org/project/ouroboros-gis/"},
+    ],
 }
 
 html_sidebars = {